+++ src/main/kotlin/com/codeland/uhc/phases/final/FinalDefault.kt
@@ -24,7 +24,7 @@ class FinalDefault : Phase() {
 		}
 	}
 
-	override fun perSecond(remainingSeconds: Long) {
+	override fun perSecond(remainingSeconds: Int) {
 
 	}
