+++ src/main/kotlin/com/codeland/uhc/phases/glowing/GlowingDefault.kt
@@ -21,7 +21,7 @@ class GlowingDefault : Phase() {
 		}
 	}
 
-	override fun perSecond(remainingSeconds: Long) {
+	override fun perSecond(remainingSeconds: Int) {
 
 	}
